requirements.txt
Django
Pillow
numpy
tensorflow
Pillow
six
object_detection